Create GUI for Atlas
