+++ src/main/java/atlas/Launcher.java
@@ -1,0 +1,16 @@
+package atlas;
+
+import javafx.application.Application;
+
+/**
+ * Class to launch the Atlas program.
+ */
+public class Launcher {
+    /**
+     * Constructs a new Atlas instance and runs it
+     * @param args Command-line arguments passed in at startup
+     */
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
