+++ src/main/java/atlas/commands/UnmarkTaskCommand.java
@@ -1,0 +1,51 @@
+package atlas.commands;
+
+import java.io.IOException;
+
+import atlas.components.Storage;
+import atlas.components.TaskList;
+import atlas.components.Ui;
+import atlas.tasks.Task;
+
+/**
+ * Command to unmark a Task in the TaskList
+ */
+public class UnmarkTaskCommand extends Command {
+    private final int idx;
+
+    /**
+     * Constructs a MarkTaskCommand
+     * @param idx Index of task to mark
+     */
+    public UnmarkTaskCommand(int idx) {
+        this.idx = idx;
+    }
+
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        try {
+            Task t = taskList.unmarkTask(idx);
+            storage.save(taskList);
+            ui.printToScreen("OK, I've marked this task as not done yet:\n"
+                    + "\t" + t);
+        } catch (IndexOutOfBoundsException e) {
+            ui.showError(String.format("%d is not a valid index! Unable to unmark task.", idx + 1));
+        } catch (IOException e) {
+            ui.showError(e.getMessage());
+        }
+    }
+
+    @Override
+    public String execute(TaskList taskList, Storage storage) {
+        try {
+            Task t = taskList.unmarkTask(idx);
+            storage.save(taskList);
+            return "OK, I've marked this task as not done yet:\n"
+                    + "\t" + t;
+        } catch (IndexOutOfBoundsException e) {
+            return String.format("%d is not a valid index! Unable to unmark task.", idx + 1);
+        } catch (IOException e) {
+            return e.getMessage();
+        }
+    }
+}
