+++ src/main/resources/view/MainWindow.fxml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0"
+            prefWidth="400.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1"
+            fx:controller="atlas.controllers.MainWindow">
+    <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput" prefHeight="41.0" prefWidth="324.0"
+               promptText="Enter command" AnchorPane.bottomAnchor="1.0">
+        <font>
+        <Font name="System Font" size="16.0"/>
+        </font>
+    </TextField>
+    <Button fx:id="sendButton" layoutX="328.0" layoutY="558.0" mnemonicParsing="false" onAction="#handleUserInput"
+            prefHeight="41.0" prefWidth="69.0" style="-fx-background-radius: 10;" text="Send" textAlignment="CENTER">
+        <font>
+        <Font size="16.0"/>
+        </font>
+    </Button>
+    <ScrollPane fx:id="scrollPane" hbarPolicy="NEVER" hvalue="1.0" prefHeight="557.0" prefWidth="400.0"
+                vvalue="1.0">
+        <VBox fx:id="dialogContainer" prefWidth="382.0"/>
+    </ScrollPane>
+</AnchorPane>
